+++ lib/net/message-socket.js
@@ -182,3 +182,4 @@ class MessageSocket extends EventEmitter {
 }
 
 module.exports = MessageSocket;
+module.exports.MessageSocket = MessageSocket;
