+++ lib/utils/test.d.ts
@@ -4,5 +4,8 @@
  * The full license text can be found in LICENSE.txt file and
  * on the Eclipse official site (https://www.eclipse.org/legal/epl-v10.html).
  */
+import * as P from 'bluebird';
 
 export function logStub(): Object;
+
+export function waitForCondition(condition: any, deadline?: number, checkPeriod?: number): P<any>;
