+++ package.json
@@ -36,6 +36,8 @@
   },
   "homepage": "https://github.com/untu/comedy#readme",
   "devDependencies": {
+    "@types/bson": "^1.0.11",
+    "@types/supertest": "^2.0.7",
     "chai": "^3.5.0",
     "chai-like": "^0.1.10",
     "eslint": "4.3.0",
@@ -53,8 +55,8 @@
     "np": "^2.13.1",
     "nyc": "^11.0.3",
     "supertest": "^3.0.0",
-    "tslint": "5.10.0",
     "ts-node": "8.0.2",
+    "tslint": "5.10.0",
     "typescript": "3.0.3"
   },
   "dependencies": {
