+++ test-resources/ts-resources/types/is-running.d.ts
@@ -1,0 +1,9 @@
+// Type definitions for is-running 2.1.0
+// Project: https://github.com/nisaacson/is-running
+// Definitions by: noctifer20 <https://github.com/noctifer20>,
+
+declare module 'is-running' {
+  function isRunning(pid: number): boolean;
+  namespace isRunning {}
+  export = isRunning;
+}
